Reorganise files into common package
