+++ src/main/java/Victor/Commands/MarkCommand.java
@@ -1,17 +1,13 @@
-package Commands;
+package Victor.Commands;
 
-import Messages.ReturnMessage;
-import Tasks.Task;
+import Victor.Messages.ReturnMessage;
 
-import java.io.FileWriter;
-import java.io.IOException;
 import java.nio.file.Path;
-import java.util.ArrayList;
 
-public class DeleteCommand extends Command {
+public class MarkCommand extends Command {
     private String taskNumber;
 
-    public DeleteCommand(String[] additionalInput) {
+    public MarkCommand(String[] additionalInput) {
         super(additionalInput);
         if (additionalInput.length == 1) {
             this.taskNumber = "-1";
@@ -20,13 +16,17 @@ public class DeleteCommand extends Command {
         }
     }
 
+    public MarkCommand(int task) {
+        super(new String[] {});
+        this.taskNumber = String.valueOf(task);
+    }
+
     @Override
     public ReturnMessage execute() {
         try {
-            return new ReturnMessage(super.taskList.deleteTask(Integer.parseInt(taskNumber)));
+            return new ReturnMessage(super.taskList.markDone(Integer.parseInt(taskNumber)));
         } catch (NumberFormatException e) {
-            return new ReturnMessage("  ~  Sorry, I don't think you entered a number for which" +
-                    " task to delete!");
+            return new ReturnMessage("  ~  I don't think there's a task with that number!");
         }
     }
