+++ src/main/java/Victor/Parser/Parser.java
@@ -1,6 +1,14 @@
-package Parser;
+package Victor.Parser;
 
-import Commands.*;
+import Victor.Commands.Command;
+import Victor.Commands.DeadlineCommand;
+import Victor.Commands.DeleteCommand;
+import Victor.Commands.EventCommand;
+import Victor.Commands.ExitCommand;
+import Victor.Commands.ListCommand;
+import Victor.Commands.MarkCommand;
+import Victor.Commands.ToDoCommand;
+import Victor.Commands.UnmarkCommand;
 
 public class Parser {
     public Command parseInput(String input) {
