+++ src/main/java/Victor/Tasks/Event.java
@@ -1,4 +1,4 @@
-package Tasks;
+package Victor.Tasks;
 
 import java.io.FileWriter;
 import java.io.IOException;
