+++ src/main/java/Victor/Tasks/ToDo.java
@@ -1,6 +1,6 @@
-package Tasks;
+package Victor.Tasks;
 
-import Tasks.Task;
+import Victor.Tasks.Task;
 
 import java.io.FileWriter;
 import java.io.IOException;
